backend/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.3.2

# AWS SDK
boto3==1.34.34
botocore==1.34.34

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
httpx==0.26.0
python-dateutil==2.8.2
pillow==10.2.0
numpy==1.26.3

# Async support
aiofiles==23.2.1
asyncpg==0.29.0

# Validation
email-validator==2.1.0.post1

# Monitoring & Logging
prometheus-client==0.19.0

# Retry and Resilience
tenacity==8.2.3

# AI/ML Libraries for Volume Estimation
torch==2.1.2
torchvision==0.16.2
timm==0.9.12
opencv-python==4.9.0.80
scikit-image==0.22.0
einops==0.7.0

# YOLO for scale detection
ultralytics==8.1.0
